wrap product ui state

+++ core/src/main/java/org/the_chance/honeymart/domain/model/Product.kt
@@ -7,14 +7,14 @@ import org.the_chance.honeymart.data.source.remote.models.ProductDto
  */
 
 data class Product(
-    val id: Int,
+    val id: Long,
     val name: String,
     val quantity: String,
     val price: Double,
 )
 
   fun ProductDto.asProduct() = Product(
-    id = id.toInt(),
+    id = id,
     name = name,
     quantity = quantity,
     price = price,
